Fixed a bug in server.py.
Fixed a bug in server.py. At line 293, we accidentally repeated the line "self.aggregateAfterAccumulation(sampled_client_indices, mixing_coefficients) ". After careful checking, the bug originated from a mistake in code sorting. It does not appear in our experimental works. That is, our experimental results are unrelated to the bug.